requirements.txt
aiofiles==24.1.0
aiogram==3.23.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiomysql==0.3.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
asyncio==4.0.0
attrs==25.4.0
beautifulsoup4==4.14.2
bitarray==3.8.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
ckzg==2.1.5
cryptography==46.0.3
cytoolz==1.1.0
decorator==5.2.1
distro==1.9.0
dotenv==0.9.9
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
executing==2.2.1
filelock==3.20.0
frozenlist==1.8.0
getenv==0.2.0
google-api-core==2.28.1
google-auth==2.43.0
google-genai==1.54.0
googleapis-common-protos==1.72.0
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipython==8.37.0
jedi==0.19.2
jiter==0.12.0
lxml==6.0.2
magic-filter==1.0.12
matplotlib-inline==0.2.1
multidict==6.7.0
mysql==0.0.3
mysqlclient==2.2.7
ntscraper==0.4.0
oauthlib==3.3.1
openai==2.14.0
parsimonious==0.10.0
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.10
pydantic_core==2.33.2
Pygments==2.19.2
PyMySQL==1.1.2
PySocks==1.7.1
python-dotenv==1.2.1
pyunormalize==17.0.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rlp==4.1.0
rsa==4.9.1
sniffio==1.3.1
snscrape==0.7.0.20230622
soupsieve==2.8
stack-data==0.6.3
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.3.3
toolz==1.1.0
tqdm==4.67.1
traitlets==5.14.3
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.14
web3==7.14.0
websockets==15.0.1
yarl==1.22.0
